fix: token reset error (#5)
