+++ src/app/(causw)/home/page.tsx
@@ -27,7 +27,7 @@ const cardsEntities: HomeCardProps[] = [
     title: "Choosing a locker",
     subtitle: "사물함 예약하기",
     icon: "/homeIcons/locker.png",
-    href: "/auth/lockers", // FIXME: Change to actual locker page
+    href: "/lockers", // FIXME: Change to actual locker page
     bgColor: "bg-[rgba(118,198,209,1)]",
   },
 ];
